gui_requirements.txt
# GUI Requirements for unbeQuant

# Web-based GUI (Dash) - original implementation
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.17.0

# Tkinter GUI - new implementation for large heatmaps
matplotlib>=3.7.0
graphviz>=0.20.0

# Common requirements for both GUIs
pandas>=2.0.0
numpy>=1.26.4
Pillow>=10.0.0
h5py>=3.10.0
pyopenms>=3.1.0
scipy>=1.11.0
python-igraph>=0.11.0
networkx>=3.1
